+++ backend/config/logbook.js
@@ -1,5 +1,6 @@
 
 module.exports = {
-    // url: "http://elog.csns.ihep.ac.cn/",
-    url: "https://elog.psi.ch/elogs/"
+    // Do not include trailing '/'
+    // url: "http://elog.csns.ihep.ac.cn",
+    url: "https://elog.psi.ch/elogs"
 }