+++ frontend/src/config/configuration.js
@@ -1,7 +1,9 @@
 module.exports = {
+    localStorageUser: "elogMobileUser",
+    localStorageOptions: "elogMobileOptions",
+
     // Development
     serverPath: "http://localhost:3001",
-    // serverPath: "http://10.1.44.253:3001",
 
     // Production
     // serverPath: "http://10.1.236.126:3001",