+++ frontend/src/router.js
@@ -5,19 +5,22 @@ const routes = [
 	{
 		path: '/',
 		name: 'home',
-		meta: { keepAlive: true },
 		component: () => import('./components/Home.vue')
 	},
 	{
 		path: '/logbook/:logbook',
 		name: 'logbook',
-		meta: { keepAlive: true },
+		meta: { keepAlive: true }, // Enable component cache
 		component: () => import('./components/Logbook.vue')
+	},
+	{
+		path: '/logbook/:logbook/:log',
+		name: 'logedit',
+		component: () => import('./components/LogEdit.vue')
 	},
 	{
 		path: '/log/:logbook/:id',
 		name: 'log',
-		meta: { keepAlive: true },
 		component: () => import('./components/Log.vue')
 	},
 	{
@@ -36,6 +39,11 @@ const router = createRouter({
 	history: createWebHistory(process.env.BASE_URL),
     routes: routes
 });
+
+// Add previous route
+// router.afterEach((to, from) => {
+// 	to.from = from;
+// });
 
 
 export default router;