+++ frontend/src/store/index.js
@@ -1,0 +1,9 @@
+import { createStore } from 'vuex'
+import { authentication } from './authentication_store';
+
+
+export const store = createStore({
+    modules: {
+        authentication,
+    }
+})