+++ frontend/CHANGELOG.md
@@ -1,3 +1,28 @@
 # Changelog
 
-## 1.0.0
+## Version history of elog-mobile
+
+### v1.0
+
+#### v1.0.0
+
+**Main changes:**
+
+- Provides function of user login
+- Provides functions of creating log, editing log and deleting log 
+
+### v0.1
+
+#### v0.1.0
+
+**Main changes:**
+
+- Updates versions of Node.js, Vue.js and PrimeVue
+
+### v0.0
+
+#### v0.0.0
+
+**Initial version:**
+
+- Provides the function of reading logs